+++ recipes/_user.rb
@@ -2,6 +2,20 @@
 # Cookbook Name:: dop_base
 # Recipe:: _user
 #
+# Copyright 2015, Achim Rosenhagen
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
 
 include_recipe 'user'
 include_recipe 'sudo'
@@ -26,14 +40,42 @@ template "#{node['users']['deploy']['home']}/wrap-ssh4git.sh" do
   mode 0700
 end
 
-ssh_config '*' do
-  user node['users']['deploy']['name']
-  options StrictHostKeyChecking: 'no',
-          UserKnownHostsFile: '/etc/ssh/ssh_known_hosts',
-          ControlPersist: 'yes',
-          ControlMaster: 'yes',
-          ControlPath: '/tmp/%r@%h:%p',
-          NoHostAuthenticationForLocalhost: 'yes'
+if ubuntu_after_saucy? || debian_after_wheezy?
+  ssh_config '*' do
+    user node['users']['deploy']['name']
+    options StrictHostKeyChecking: 'yes',
+            UserKnownHostsFile: '/etc/ssh/ssh_known_hosts',
+            ControlPersist: 'yes',
+            ControlMaster: 'yes',
+            ControlPath: '/tmp/%r@%h:%p',
+            NoHostAuthenticationForLocalhost: 'yes',
+            KexAlgorithms: 'curve25519-sha256@libssh.org,diffie-hellman-group-exchange-sha256',
+            Protocol: '2',
+            HostKey: '/etc/ssh/ssh_host_ed25519_key',
+            HostKey: '/etc/ssh/ssh_host_rsa_key',
+            PasswordAuthentication: 'no',
+            PubkeyAuthentication: 'yes',
+            Ciphers: 'chacha20-poly1305@openssh.com,aes256-gcm@openssh.com,aes128-gcm@openssh.com,aes256-ctr,aes192-ctr,aes128-ctr',
+            MACs: 'hmac-sha2-512-etm@openssh.com,hmac-sha2-256-etm@openssh.com,hmac-ripemd160-etm@openssh.com,umac-128-etm@openssh.com,hmac-sha2-512,hmac-sha2-256,hmac-ripemd160,umac-128@openssh.com'
+  end
+else
+  ssh_config '*' do
+    user node['users']['deploy']['name']
+    options StrictHostKeyChecking: 'yes',
+            UserKnownHostsFile: '/etc/ssh/ssh_known_hosts',
+            ControlPersist: 'yes',
+            ControlMaster: 'yes',
+            ControlPath: '/tmp/%r@%h:%p',
+            NoHostAuthenticationForLocalhost: 'yes',
+            KexAlgorithms: 'diffie-hellman-group-exchange-sha256',
+            Protocol: '2',
+            HostKey: '/etc/ssh/ssh_host_ed25519_key',
+            HostKey: '/etc/ssh/ssh_host_rsa_key',
+            PasswordAuthentication: 'no',
+            PubkeyAuthentication: 'yes',
+            Ciphers: 'aes256-ctr,aes192-ctr,aes128-ctr',
+            MACs: 'hmac-sha2-512,hmac-sha2-256,hmac-ripemd160'
+  end
 end
 
 cookbook_file "#{node['users']['deploy']['home']}/.gemrc" do