+++ test/integration/default/serverspec/locale_spec.rb
@@ -1,6 +1,4 @@
-require 'serverspec'
-set :backend, :exec
-set :path, '/sbin:/usr/sbin:/usr/local/sbin:$PATH'
+require_relative '../../../kitchen/data/spec_helper'
 
 describe command('echo $LC_ALL'), if: os[:family] == 'debian' do
   its(:stdout) { should match 'en_US.UTF-8' }